+++ test-setup.ts
@@ -1,2 +1,10 @@
 // oxlint-disable-next-line import/no-unassigned-import
 import "@testing-library/jest-dom";
+
+class ResizeObserver {
+  observe() {}
+  unobserve() {}
+  disconnect() {}
+}
+
+globalThis.ResizeObserver = ResizeObserver;